+++ packages/analytics-js/__tests__/components/configManager/commonUtil.test.ts
@@ -110,7 +110,7 @@ describe('Config Manager Common Utilities', () => {
 
       updateReportingState(mockSourceConfig, mockLogger);
 
-      expect(state.reporting.isErrorReportingEnabled.value).toBe(true);
+      // expect(state.reporting.isErrorReportingEnabled.value).toBe(true); // TODO: uncomment this line when error reporting is enabled
       expect(state.reporting.isMetricsReportingEnabled.value).toBe(true);
       expect(mockLogger.warn).not.toHaveBeenCalled();
     });
@@ -133,7 +133,7 @@ describe('Config Manager Common Utilities', () => {
 
       updateReportingState(mockSourceConfig, mockLogger);
 
-      expect(state.reporting.isErrorReportingEnabled.value).toBe(true);
+      // expect(state.reporting.isErrorReportingEnabled.value).toBe(true); // TODO: uncomment this line when error reporting is enabled
       expect(state.reporting.isMetricsReportingEnabled.value).toBe(true);
       expect(mockLogger.warn).not.toHaveBeenCalled();
     });