+++ packages/analytics-js-plugins/CHANGELOG.md
@@ -2,6 +2,27 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.4.6](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.4.5...@rudderstack/analytics-js-plugins@3.4.6) (2024-07-24)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.4.5`
+* `@rudderstack/analytics-js-cookies` updated to version `3.4.5`
+## [3.4.5](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.4.4...@rudderstack/analytics-js-plugins@3.4.5) (2024-07-23)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.4.4`
+## [3.4.4](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.4.3...@rudderstack/analytics-js-plugins@3.4.4) (2024-07-23)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.4.3`
+## [3.4.3](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.4.2...@rudderstack/analytics-js-plugins@3.4.3) (2024-07-23)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.4.2`
 ## [3.4.2](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.4.1...@rudderstack/analytics-js-plugins@3.4.2) (2024-07-22)
 
 ### Dependency Updates