+++ packages/analytics-js-service-worker/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.1.6](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.1.5...@rudderstack/analytics-js-service-worker@3.1.6) (2024-07-24)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.1.5`
 ## [3.1.5](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.1.4...@rudderstack/analytics-js-service-worker@3.1.5) (2024-07-19)
 
 ### Dependency Updates