+++ packages/loading-scripts/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.0.23](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.22...@rudderstack/analytics-js-loading-scripts@3.0.23) (2024-07-24)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.0.22`
+## [3.0.22](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.21...@rudderstack/analytics-js-loading-scripts@3.0.22) (2024-07-23)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.0.21`
+## [3.0.21](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.20...@rudderstack/analytics-js-loading-scripts@3.0.21) (2024-07-23)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.0.20`
+## [3.0.20](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.19...@rudderstack/analytics-js-loading-scripts@3.0.20) (2024-07-23)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.0.19`
 ## [3.0.19](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.18...@rudderstack/analytics-js-loading-scripts@3.0.19) (2024-07-22)
 
 ### Dependency Updates