+++ packages/sanity-suite/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.1.13](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-sanity-suite@3.1.12...@rudderstack/analytics-js-sanity-suite@3.1.13) (2024-07-24)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.1.12`
+## [3.1.12](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-sanity-suite@3.1.11...@rudderstack/analytics-js-sanity-suite@3.1.12) (2024-07-23)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.1.11`
+## [3.1.11](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-sanity-suite@3.1.10...@rudderstack/analytics-js-sanity-suite@3.1.11) (2024-07-23)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.1.10`
+## [3.1.10](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-sanity-suite@3.1.9...@rudderstack/analytics-js-sanity-suite@3.1.10) (2024-07-23)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.1.9`
 ## [3.1.9](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-sanity-suite@3.1.8...@rudderstack/analytics-js-sanity-suite@3.1.9) (2024-07-22)
 
 ### Dependency Updates